chore: move common code
