+++ src/common.js
@@ -1,4 +1,20 @@
-const isPlainObj = require('is-plain-obj')
+function addSuccess(result, object) {
+  return { ...result, success: [...result.success, object] }
+}
+
+function addError(result, object) {
+  return { ...result, errors: [...result.errors, object] }
+}
+
+function isInvalidSource(redirect) {
+  return redirect.path.match(/^\/\.netlify/)
+}
+
+function isProxy(redirect) {
+  return Boolean(redirect.proxy || (/^https?:\/\//.test(redirect.to) && redirect.status === 200))
+}
+
+const FULL_URL_MATCHER = /^(https?):\/\/(.+)$/
 
 let URLclass = null
 
@@ -12,72 +28,6 @@ function parseURL(url) {
   // eslint-disable-next-line node/global-require
   URLclass = URLclass || require('url')
   return new URLclass.URL(url)
-}
-
-function splatForwardRule(path, obj, dest) {
-  return path.match(/\/\*$/) && dest == null && obj.status && obj.status >= 200 && obj.status < 300 && obj.force
-}
-
-function fetch(obj, options) {
-  for (const option in options) {
-    if (Object.prototype.hasOwnProperty.call(obj, options[option])) {
-      return obj[options[option]]
-    }
-  }
-  return null
-}
-
-function redirectMatch(obj) {
-  const origin = fetch(obj, ['from', 'origin'])
-  const redirect = origin && FULL_URL_MATCHER.test(origin) ? parseFullOrigin(origin) : { path: origin }
-  if (redirect == null || (redirect.path == null && redirect.host == null)) {
-    return null
-  }
-
-  const dest = fetch(obj, ['to', 'destination'])
-  redirect.to = splatForwardRule(redirect.path, obj, dest) ? redirect.path.replace(/\/\*$/, '/:splat') : dest
-
-  if (redirect.to == null) {
-    return null
-  }
-
-  redirect.params = fetch(obj, ['query', 'params', 'parameters'])
-  redirect.status = fetch(obj, ['status'])
-  redirect.force = fetch(obj, ['force'])
-  redirect.conditions = fetch(obj, ['conditions'])
-  redirect.headers = fetch(obj, ['headers'])
-  redirect.signed = fetch(obj, ['sign', 'signing', 'signed'])
-
-  Object.keys(redirect).forEach((key) => {
-    if (redirect[key] === null) {
-      delete redirect[key]
-    }
-  })
-
-  if (redirect.headers && !isPlainObj(redirect.headers)) {
-    return null
-  }
-
-  return redirect
-}
-
-function addSuccess(result, object) {
-  return { ...result, success: [...result.success, object] }
-}
-
-function addError(result, object) {
-  return { ...result, errors: [...result.errors, object] }
-}
-
-const FULL_URL_MATCHER = /^(https?):\/\/(.+)$/
-const FORWARD_STATUS_MATCHER = /^2\d\d!?$/
-
-function isInvalidSource(redirect) {
-  return redirect.path.match(/^\/\.netlify/)
-}
-
-function isProxy(redirect) {
-  return Boolean(redirect.proxy || (/^https?:\/\//.test(redirect.to) && redirect.status === 200))
 }
 
 function parseFullOrigin(origin) {
@@ -96,13 +46,10 @@ function parseFullOrigin(origin) {
 }
 
 module.exports = {
-  splatForwardRule,
-  redirectMatch,
   addSuccess,
   addError,
-  FULL_URL_MATCHER,
-  FORWARD_STATUS_MATCHER,
   isInvalidSource,
   isProxy,
+  FULL_URL_MATCHER,
   parseFullOrigin,
 }
